working aos workflow
