+++ .eslintrc.json
@@ -60,6 +60,7 @@
           "**/build-tools/**",
           "**/handler/**/*.ts",
           "**/utilities/**/*.ts",
+          "src/common/lambda/*.ts",
           ".projenrc.ts",
           "projenrc/**/*.ts"
         ],
