+++ package.json
@@ -53,6 +53,7 @@
   "devDependencies": {
     "@aws-cdk/integ-runner": "latest",
     "@aws-cdk/integ-tests-alpha": "latest",
+    "@aws-crypto/sha256-js": "5.2.0",
     "@aws-sdk/client-cloudformation": "3.600.0",
     "@aws-sdk/client-dynamodb": "3.600.0",
     "@aws-sdk/client-ec2": "^3.760.0",
@@ -63,16 +64,21 @@
     "@aws-sdk/client-s3": "3.600.0",
     "@aws-sdk/client-secrets-manager": "3.600.0",
     "@aws-sdk/client-ssm": "3.600.0",
+    "@aws-sdk/client-sts": "3.600.0",
     "@aws-sdk/client-wafv2": "^3.758.0",
+    "@aws-sdk/credential-provider-node": "3.600.0",
     "@aws-sdk/lib-dynamodb": "3.600.0",
     "@aws-sdk/s3-request-presigner": "3.600.0",
     "@codegenie/serverless-express": "4.16.0",
+    "@smithy/signature-v4": "5.0.1",
     "@smithy/types": "^3.7.2",
     "@smithy/util-stream-node": "4.0.2",
+    "@types/adm-zip": "0.5.7",
     "@types/aws-lambda": "8.10.141",
     "@types/express": "5.0.0",
     "@types/glob": "^8.1.0",
     "@types/jest": "^29.5.14",
+    "@types/js-yaml": "^4.0.9",
     "@types/mime": "3.0.0",
     "@types/morgan": "1.9.9",
     "@types/node": "^18",
@@ -80,6 +86,7 @@
     "@types/uuid": "^10.0.0",
     "@typescript-eslint/eslint-plugin": "^8",
     "@typescript-eslint/parser": "^8",
+    "adm-zip": "0.5.16",
     "aws-cdk-lib": "2.177.0",
     "aws-sdk-client-mock": "^4.1.0",
     "aws-sdk-client-mock-jest": "^4.1.0",
