+++ src/constructs/opensearch-workflow/handler/types/resource-response.ts
