+++ src/types/destructive-operation.ts
@@ -1,0 +1,8 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+export enum DestructiveOperation {
+    UPDATE = 'UPDATE',
+    DELETE = 'DELETE',
+    ALL = 'ALL'
+}
