+++ src/types/index.ts
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 export * from './aws-managed-policy';
+export * from './destructive-operation';
 export * from './lambda-configuration';
 export * from './ec2-instance-type';
 export * from './sagemaker-notebook-instance-type';
