+++ utilities/projen/lib/deps.ts
@@ -22,13 +22,19 @@ export const configureDependencies = (project: CdklabsTypeScriptProject) => {
         `@aws-sdk/client-cloudformation@${sdkVersion}`,
         `@aws-sdk/client-s3@${sdkVersion}`,
         `@aws-sdk/s3-request-presigner@${sdkVersion}`,
+        `@aws-sdk/client-sts@${sdkVersion}`,
+        `@aws-sdk/credential-provider-node@${sdkVersion}`,
+        `@aws-crypto/sha256-js@5.2.0`,
         '@codegenie/serverless-express@4.16.0',
+        '@smithy/signature-v4@5.0.1',
         '@smithy/types@^3.7.2',
         '@smithy/util-stream-node@4.0.2',
+        '@types/adm-zip@0.5.7',
         '@types/aws-lambda@8.10.141',
         '@types/express@5.0.0',
         '@types/mime@3.0.0',
         '@types/morgan@1.9.9',
+        'adm-zip@0.5.16',
         'axios@1.7.7',
         'express@5.0.1',
         'express-async-handler@1.2.0',
@@ -45,6 +51,7 @@ export const configureDependencies = (project: CdklabsTypeScriptProject) => {
         '@types/glob',
         '@types/supertest',
         '@types/uuid',
+        '@types/js-yaml',
         'aws-sdk-client-mock',
         'aws-sdk-client-mock-jest',
         'cdk-nag@2.34.0',
